StringAvatar をコンポーネントに切り出し #29
